Rename pipe_works to pipe_run, move PipeRunParams in there rename dry_run.py
