+++ pipelex/cli/commands/validate_cmd.py
@@ -8,7 +8,7 @@ import typer
 from pipelex import log
 from pipelex.cli.commands.common import is_pipelex_libraries_folder
 from pipelex.hub import get_pipeline_tracker, get_pipes, get_required_pipe
-from pipelex.pipe_works.pipe_dry import dry_run_pipe, dry_run_pipes
+from pipelex.pipe_run.dry_run import dry_run_pipe, dry_run_pipes
 from pipelex.pipelex import Pipelex
