+++ pipelex/client/pipeline_request_factory.py
@@ -3,7 +3,7 @@ from typing import Any
 from pipelex.client.api_serializer import ApiSerializer
 from pipelex.client.protocol import COMPACT_MEMORY_KEY, PipelineRequest
 from pipelex.core.memory.working_memory import WorkingMemory
-from pipelex.core.pipes.pipe_run_params import PipeOutputMultiplicity
+from pipelex.pipe_run.pipe_run_params import PipeOutputMultiplicity
 
 
 class PipelineRequestFactory:
