+++ pipelex/pipe_controllers/sub_pipe_factory.py
@@ -1,6 +1,6 @@
-from pipelex.core.pipes.pipe_run_params import BatchParams, make_output_multiplicity
 from pipelex.pipe_controllers.sub_pipe import SubPipe
 from pipelex.pipe_controllers.sub_pipe_blueprint import SubPipeBlueprint
+from pipelex.pipe_run.pipe_run_params import BatchParams, make_output_multiplicity
 
 
 class SubPipeFactory:
