+++ pipelex/pipe_operators/llm/pipe_llm.py
@@ -20,12 +20,6 @@ from pipelex.core.memory.working_memory import WorkingMemory
 from pipelex.core.pipes.input_requirements import InputRequirements
 from pipelex.core.pipes.input_requirements_factory import InputRequirementsFactory
 from pipelex.core.pipes.pipe_output import PipeOutput
-from pipelex.core.pipes.pipe_run_params import (
-    PipeOutputMultiplicity,
-    PipeRunParamKey,
-    PipeRunParams,
-    output_multiplicity_to_apply,
-)
 from pipelex.core.stuffs.list_content import ListContent
 from pipelex.core.stuffs.stuff_content import StuffContent
 from pipelex.core.stuffs.stuff_factory import StuffFactory
@@ -49,6 +43,12 @@ from pipelex.hub import (
 )
 from pipelex.pipe_operators.llm.pipe_llm_blueprint import StructuringMethod
 from pipelex.pipe_operators.pipe_operator import PipeOperator
+from pipelex.pipe_run.pipe_run_params import (
+    PipeOutputMultiplicity,
+    PipeRunParamKey,
+    PipeRunParams,
+    output_multiplicity_to_apply,
+)
 from pipelex.pipeline.job_metadata import JobMetadata
 from pipelex.tools.templating.jinja2_template_category import Jinja2TemplateCategory
 from pipelex.tools.typing.structure_printer import StructurePrinter
