+++ pipelex/pipe_run/pipe_job_factory.py
@@ -1,9 +1,9 @@
 from pipelex.core.memory.working_memory import WorkingMemory
 from pipelex.core.memory.working_memory_factory import WorkingMemoryFactory
 from pipelex.core.pipes.pipe_abstract import PipeAbstract
-from pipelex.core.pipes.pipe_run_params import PipeRunParams
-from pipelex.core.pipes.pipe_run_params_factory import PipeRunParamsFactory
-from pipelex.pipe_works.pipe_job import PipeJob
+from pipelex.pipe_run.pipe_job import PipeJob
+from pipelex.pipe_run.pipe_run_params import PipeRunParams
+from pipelex.pipe_run.pipe_run_params_factory import PipeRunParamsFactory
 from pipelex.pipeline.job_metadata import JobMetadata
