+++ pipelex/pipelex.py
@@ -27,8 +27,8 @@ from pipelex.exceptions import PipelexConfigError, PipelexSetupError
 from pipelex.hub import PipelexHub, get_observer_provider, set_pipelex_hub
 from pipelex.libraries.library_manager_factory import LibraryManagerFactory
 from pipelex.observer.local_observer import LocalObserver
-from pipelex.pipe_works.pipe_router import PipeRouter
-from pipelex.pipe_works.pipe_router_protocol import PipeRouterProtocol
+from pipelex.pipe_run.pipe_router import PipeRouter
+from pipelex.pipe_run.pipe_router_protocol import PipeRouterProtocol
 from pipelex.pipeline.activity.activity_manager import ActivityManager
 from pipelex.pipeline.activity.activity_manager_protocol import (
     ActivityManagerNoOp,
