+++ tests/integration/pipelex/pipes/operator/pipe_imgg/test_pipe_img_gen.py
@@ -4,12 +4,12 @@ import pytest
 
 from pipelex import pretty_print
 from pipelex.core.concepts.concept_native import NativeConceptEnum
-from pipelex.core.pipes.pipe_run_params import PipeRunMode
-from pipelex.core.pipes.pipe_run_params_factory import PipeRunParamsFactory
 from pipelex.hub import get_pipe_router
 from pipelex.pipe_operators.img_gen.pipe_img_gen_blueprint import PipeImgGenBlueprint
 from pipelex.pipe_operators.img_gen.pipe_img_gen_factory import PipeImgGenFactory
-from pipelex.pipe_works.pipe_job_factory import PipeJobFactory
+from pipelex.pipe_run.pipe_job_factory import PipeJobFactory
+from pipelex.pipe_run.pipe_run_params import PipeRunMode
+from pipelex.pipe_run.pipe_run_params_factory import PipeRunParamsFactory
 from tests.integration.pipelex.test_data import ImageGenTestCases
 
 if TYPE_CHECKING:
